requirements.txt
watchdog>=4.0
anthropic>=0.40
openpyxl>=3.1
Pillow>=10.0
pypdfium2>=5.0
pystray>=0.19
winotify>=1.1
